Extract shared service data models
